requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
python-dotenv==1.2.2
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
twilio==9.10.2
pydantic==2.10.4
python-multipart==0.0.20
requests==2.32.5
websockets==16.0
email-validator==2.3.0
httpx==0.28.1
google-api-python-client==2.169.0
google-auth==2.40.2
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
stripe==12.0.1
python-dateutil==2.9.0.post0
phonenumbers==9.0.4
orjson==3.10.18